requirements.txt
streamlit
pytesseract
pdf2image
Pillow
pdfplumber
opencv-python-headless
google-genai
python-dotenv
python-dateutil
streamlit-google-auth
openpyxl
pandas
sentence-transformers
torch
pillow
